refactor: move comment conversion to common
