+++ draft-api/src/main/scala/no/ndla/draftapi/service/ConverterService.scala
@@ -11,21 +11,12 @@ package no.ndla.draftapi.service
 import cats.implicits.*
 import com.typesafe.scalalogging.StrictLogging
 import no.ndla.common.configuration.Constants.EmbedTagName
-import no.ndla.common.errors.{ValidationException, ValidationMessage}
+import no.ndla.common.converter.CommonConverter
+import no.ndla.common.errors.ValidationException
 import no.ndla.common.implicits.TryQuestionMark
-import no.ndla.common.model.api.{
-  CommentDTO,
-  Delete,
-  DisclaimerDTO,
-  DraftCopyrightDTO,
-  Missing,
-  NewCommentDTO,
-  ResponsibleDTO,
-  UpdateWith,
-  UpdatedCommentDTO
-}
-import no.ndla.common.model.domain.{ArticleContent, Comment, Priority, Responsible}
-import no.ndla.common.model.domain.draft.DraftStatus.{IMPORTED, PLANNED}
+import no.ndla.common.model.api.{Delete, DisclaimerDTO, DraftCopyrightDTO, Missing, ResponsibleDTO, UpdateWith}
+import no.ndla.common.model.domain.{ArticleContent, Priority, Responsible}
+import no.ndla.common.model.domain.draft.DraftStatus.PLANNED
 import no.ndla.common.model.domain.draft.{Draft, DraftStatus}
 import no.ndla.common.model.domain.language.OptLanguageFields
 import no.ndla.common.model.{RelatedContentLink, api as commonApi, domain as common}
@@ -50,7 +41,8 @@ import scala.jdk.CollectionConverters.*
 import scala.util.{Failure, Success, Try}
 
 trait ConverterService {
-  this: Clock & DraftRepository & ArticleApiClient & StateTransitionRules & WriteService & UUIDUtil & Props =>
+  this: Clock & DraftRepository & ArticleApiClient & StateTransitionRules & WriteService & UUIDUtil & CommonConverter &
+    Props =>
   val converterService: ConverterService
 
   class ConverterService extends StrictLogging {
@@ -123,52 +115,13 @@ trait ConverterService {
           revisionMeta = revisionMeta,
           responsible = responsible,
           slug = newArticle.slug,
-          comments = newCommentToDomain(newArticle.comments.getOrElse(List.empty)),
+          comments = newArticle.comments
+            .map(comments => comments.map(CommonConverter.newCommentApiToDomain))
+            .getOrElse(List.empty),
           priority = priority,
           started = false,
           qualityEvaluation = qualityEvaluationToDomain(newArticle.qualityEvaluation),
           disclaimer = domainDisclaimer
-        )
-      )
-    }
-
-    private[service] def updatedCommentToDomain(
-        updatedComments: List[UpdatedCommentDTO],
-        existingComments: Seq[Comment]
-    ): Seq[Comment] = {
-      updatedComments.map(updatedComment => {
-        existingComments.find(cc => updatedComment.id.contains(cc.id.toString)) match {
-          case Some(existingComment) =>
-            val isContentChanged = updatedComment.content != existingComment.content
-            val newUpdated       = if (isContentChanged) clock.now() else existingComment.updated
-            existingComment.copy(
-              updated = newUpdated,
-              content = updatedComment.content,
-              isOpen = updatedComment.isOpen.getOrElse(true),
-              solved = updatedComment.solved.getOrElse(false)
-            )
-          case None =>
-            Comment(
-              id = uuidUtil.randomUUID(),
-              created = clock.now(),
-              updated = clock.now(),
-              content = updatedComment.content,
-              isOpen = updatedComment.isOpen.getOrElse(true),
-              solved = updatedComment.solved.getOrElse(false)
-            )
-        }
-      })
-    }
-
-    private[service] def newCommentToDomain(newComment: List[NewCommentDTO]): Seq[Comment] = {
-      newComment.map(comment =>
-        Comment(
-          id = UUID.randomUUID(),
-          created = clock.now(),
-          updated = clock.now(),
-          content = comment.content,
-          isOpen = comment.isOpen.getOrElse(true),
-          solved = false
         )
       )
     }
@@ -185,37 +138,6 @@ trait ConverterService {
         introduction = article.introduction.sorted,
         metaImage = article.metaImage.sorted,
         title = article.title.sorted
-      )
-    }
-
-    private[service] def updatedCommentToDomainNullDocument(
-        updatedComments: List[UpdatedCommentDTO]
-    ): Try[Seq[Comment]] = {
-      updatedComments.traverse(comment =>
-        comment.id match {
-          case Some(uuid) =>
-            Try(UUID.fromString(uuid)).map(uuid =>
-              Comment(
-                id = uuid,
-                created = clock.now(),
-                updated = clock.now(),
-                content = comment.content,
-                isOpen = comment.isOpen.getOrElse(true),
-                solved = comment.solved.getOrElse(false)
-              )
-            )
-          case None =>
-            Success(
-              Comment(
-                id = uuidUtil.randomUUID(),
-                created = clock.now(),
-                updated = clock.now(),
-                content = comment.content,
-                isOpen = comment.isOpen.getOrElse(true),
-                solved = comment.solved.getOrElse(false)
-              )
-            )
-        }
       )
     }
 
@@ -417,7 +339,7 @@ trait ConverterService {
             revisions = revisionMetas,
             responsible = responsible,
             slug = article.slug,
-            comments = article.comments.map(toApiComment),
+            comments = article.comments.map(CommonConverter.commentDomainToApi),
             prioritized = article.priority == Priority.Prioritized,
             priority = article.priority.entryName,
             started = article.started,
@@ -490,15 +412,6 @@ trait ConverterService {
         case Right(x) => Right(x)
       }
     }
-
-    private def toApiComment(comment: Comment): CommentDTO = model.api.CommentDTO(
-      id = comment.id.toString,
-      content = comment.content,
-      created = comment.created,
-      updated = comment.updated,
-      isOpen = comment.isOpen,
-      solved = comment.solved
-    )
 
     private def toApiQualityEvaluation(
         qualityEvaluation: Option[common.draft.QualityEvaluation]
@@ -762,7 +675,7 @@ trait ConverterService {
       val reqLibs =
         article.requiredLibraries.map(_.map(toDomainRequiredLibraries)).getOrElse(toMergeInto.requiredLibraries)
       val updatedComments = article.comments
-        .map(comments => updatedCommentToDomain(comments, toMergeInto.comments))
+        .map(comments => CommonConverter.mergeUpdatedCommentsWithExisting(comments, toMergeInto.comments))
         .getOrElse(toMergeInto.comments)
 
       val articleWithNewContent = article.copy(content = newContent)
@@ -848,83 +761,6 @@ trait ConverterService {
 
       Success(converted)
     }
-
-    def toDomainArticle(id: Long, article: api.UpdatedArticleDTO, user: TokenUser): Try[Draft] =
-      article.language match {
-        case None =>
-          val error = ValidationMessage("language", "This field must be specified when updating language fields")
-          Failure(new ValidationException(errors = Seq(error)))
-        case Some(lang) =>
-          val status      = common.Status(PLANNED, Set(IMPORTED))
-          val createdDate = clock.now()
-          val mergedNotes = article.notes.map(n => newNotes(n, user, status)) match {
-            case Some(Failure(ex))    => Failure(ex)
-            case Some(Success(notes)) => Success(notes)
-            case None                 => Success(Seq.empty)
-          }
-
-          val newMetaImage = article.metaImage match {
-            case UpdateWith(meta) => Seq(common.ArticleMetaImage(meta.id, meta.alt, lang))
-            case _                => Seq.empty
-          }
-
-          val updatedAvailability =
-            common.Availability.valueOf(article.availability).getOrElse(common.Availability.everyone)
-          val updatedRevisionMeta = article.revisionMeta.toSeq.flatMap(_.map(toDomainRevisionMeta))
-
-          val responsible = article.responsibleId match {
-            case Missing => None
-            case Delete  => None
-            case UpdateWith(responsibleId) =>
-              Some(Responsible(responsibleId = responsibleId, lastUpdated = clock.now()))
-          }
-
-          val articleType = article.articleType
-            .map(common.ArticleType.valueOfOrError)
-            .getOrElse(common.ArticleType.Standard)
-
-          val priority = common.Priority
-            .valueOfOrError(article.priority.getOrElse(Priority.Unspecified.entryName))
-            .getOrElse(common.Priority.Unspecified)
-
-          for {
-            comments <- updatedCommentToDomainNullDocument(article.comments.getOrElse(List.empty))
-            notes    <- mergedNotes
-          } yield Draft(
-            id = Some(id),
-            revision = Some(1),
-            status = status,
-            title = article.title.map(t => common.Title(t, lang)).toSeq,
-            content = article.content.map(c => common.ArticleContent(c, lang)).toSeq,
-            copyright = article.copyright.map(toDomainCopyright),
-            tags = article.tags.toSeq.map(tags => common.Tag(tags, lang)),
-            requiredLibraries = article.requiredLibraries.map(_.map(toDomainRequiredLibraries)).toSeq.flatten,
-            visualElement = article.visualElement.map(v => toDomainVisualElement(v, lang)).toSeq,
-            introduction = article.introduction.map(i => toDomainIntroduction(i, lang)).toSeq,
-            metaDescription = article.metaDescription.map(m => toDomainMetaDescription(m, lang)).toSeq,
-            metaImage = newMetaImage,
-            created = createdDate,
-            updated = createdDate,
-            published = article.published.getOrElse(clock.now()),
-            updatedBy = user.id,
-            articleType = articleType,
-            notes = notes,
-            previousVersionsNotes = Seq.empty,
-            editorLabels = article.editorLabels.getOrElse(Seq.empty),
-            grepCodes = article.grepCodes.getOrElse(Seq.empty),
-            conceptIds = article.conceptIds.getOrElse(Seq.empty),
-            availability = updatedAvailability,
-            relatedContent = article.relatedContent.map(toDomainRelatedContent).getOrElse(Seq.empty),
-            revisionMeta = updatedRevisionMeta,
-            responsible = responsible,
-            slug = article.slug,
-            comments = comments,
-            priority = priority,
-            started = false,
-            qualityEvaluation = qualityEvaluationToDomain(article.qualityEvaluation),
-            disclaimer = OptLanguageFields.fromMaybeString(article.disclaimer, lang)
-          )
-      }
 
     private[service] def buildTransitionsMap(user: TokenUser, article: Option[Draft]): Map[String, List[String]] =
       StateTransitionRules.StateTransitions.groupBy(_.from).map { case (from, to) =>
